ubuntu64-py310-requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=ubuntu64-py310-requirements.txt test-requirements.in
#
-e ./python
    # via -r test-requirements.in
astroid==2.11.5
    # via pylint
bcrypt==3.2.2
    # via paramiko
certifi==2022.5.18.1
    # via requests
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
charset-normalizer==2.0.12
    # via requests
colorama==0.4.4
    # via plush-fabric
cryptography==37.0.2
    # via
    #   paramiko
    #   secretstorage
deprecated==1.2.13
    # via pygithub
dill==0.3.5.1
    # via pylint
fabric==2.7.0
    # via
    #   patchwork
    #   plush-fabric
httplib2==0.20.4
    # via oauth2client
idna==3.3
    # via requests
importlib-metadata==4.11.4
    # via keyring
invoke==1.7.1
    # via fabric
isort==5.10.1
    # via pylint
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
keyring==23.5.1
    # via plush-fabric
lazy-object-proxy==1.7.1
    # via astroid
mccabe==0.7.0
    # via pylint
oauth2client==4.1.3
    # via plush-fabric
paramiko==2.11.0
    # via fabric
patchwork==1.0.1
    # via plush-fabric
pathlib2==2.3.7.post1
    # via fabric
platformdirs==2.5.2
    # via pylint
pyasn1==0.4.8
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via oauth2client
pycparser==2.21
    # via cffi
pygithub==1.55
    # via plush-fabric
pyjwt==2.4.0
    # via pygithub
pylint==2.14.0
    # via plush-fabric
pynacl==1.5.0
    # via
    #   paramiko
    #   pygithub
pyparsing==3.0.9
    # via httplib2
requests==2.27.1
    # via pygithub
rsa==4.8
    # via oauth2client
secretstorage==3.3.2
    # via keyring
six==1.16.0
    # via
    #   oauth2client
    #   paramiko
    #   pathlib2
tomli==2.0.1
    # via pylint
tomlkit==0.11.0
    # via pylint
urllib3==1.26.9
    # via requests
wrapt==1.14.1
    # via
    #   astroid
    #   deprecated
zipp==3.8.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
